Additional changes to incorporate Angular into Office-Addin-Taskpane
- Move app folder under taskpane folder
- Remove sample code from app.component.ts
- Import app.module into taskpane.ts
