+++ package.json
@@ -39,6 +39,7 @@
     "@types/es6-promise": "0.0.32",
     "@types/flux": "^3.1.0",
     "@types/jquery": "^3.3.22",
+    "@types/node": "^9.4.7",
     "@types/office-js": "0.0.138",
     "@types/office-runtime": "^1.0.1",
     "clean-webpack-plugin": "^1.0.0",
