+++ src/taskpane/app/app.component.ts
@@ -1,0 +1,14 @@
+import { Component } from '@angular/core';
+const template = require('./app.component.html');
+
+@Component({
+    selector: 'app-home',
+    template
+})
+export default class AppComponent {
+    welcomeMessage = 'Welcome';
+
+    async run() {
+        // Place code here.
+    }
+}
