+++ src/taskpane/app/app.module.ts
@@ -1,6 +1,5 @@
 import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
-
 import AppComponent from './app.component';
 
 @NgModule({
