+++ src/taskpane/taskpane.ts
@@ -2,7 +2,7 @@
  * Copyright (c) Microsoft Corporation. All rights reserved. Licensed under the MIT license.
  * See LICENSE in the project root for license information.
  */
-
+import AppModule from './app/app.module';
 import * as OfficeHelpers from "@microsoft/office-js-helpers";
 
 $(document).ready(() => {
